feat: #31-rename-component  - pages, components 내부 tsx 파일명 변경  - 디렉터리 이름과 tsx 파일 내부 컴포넌트 명 매칭  - name.tsx -> index.tsx 형태로 변경
