+++ src/components/dashboard/charts/bar-chart/index.tsx
