+++ src/pages/operational-statistical/index.tsx
@@ -1,15 +1,15 @@
 import { useEffect } from 'react'
 
-import { IconButton } from '@/components/common/buttons/icon-button/icon-button.tsx'
-import { Form } from '@/components/common/form/form.tsx'
-import { InputDateRange } from '@/components/common/inputs/input-date-picker/input-date-range/input-date-range.tsx'
-import { InputTimePicker } from '@/components/common/inputs/input-date-picker/input-time-picker/input-time-picker.tsx'
-import { InputSelect } from '@/components/common/inputs/input-select/input-select.tsx'
-import { TabMenu } from '@/components/dashboard/tab-menu/tab-menu.tsx'
+import { IconButton } from '@/components/common/buttons/icon-button'
+import { Form } from '@/components/common/form'
+import { InputDateRange } from '@/components/common/inputs/input-date-picker/input-date-range'
+import { InputTimePicker } from '@/components/common/inputs/input-date-picker/input-time-picker'
+import { InputSelect } from '@/components/common/inputs/input-select'
+import { TabMenu } from '@/components/dashboard/tab-menu'
 import { TabMenus, searchOptions, topFilter } from '@/configs/dashboard.ts'
 import { useSearchQuery } from '@/hooks/use-search-query.ts'
-import { ViewPage } from '@/pages/operational-statistical/view-page/view-page.tsx'
-import { WritePage } from '@/pages/operational-statistical/write-page/write-page.tsx'
+import { ViewPage } from '@/pages/operational-statistical/view-page'
+import { WritePage } from '@/pages/operational-statistical/write-page'
 import { useTabStore } from '@/stores/use-tab-store.ts'
 import { ButtonTag } from '@/types/components/common/button.ts'
 import { SearchSchema, SearchSchemaType } from '@/utils/schema/schema.ts'
