+++ src/pages/operational-statistical/view-page/index.tsx
@@ -1,8 +1,8 @@
 import { blinkCharging, chargePoint } from '@/api/mock/graph-api-data.ts'
-import { TableComponent } from '@/components/common/table/table.tsx'
-import { Accordion } from '@/components/dashboard/accordion/accordion.tsx'
-import { BarChartComp } from '@/components/dashboard/charts/bar-chart/bar-chart.tsx'
-import { DataCard } from '@/components/dashboard/count-card/count-card.tsx'
+import { TableComponent } from '@/components/common/table'
+import { Accordion } from '@/components/dashboard/accordion'
+import { BarChartComp } from '@/components/dashboard/charts/bar-chart'
+import { DataCard } from '@/components/dashboard/count-card'
 import { dataThead, tBodyStyle } from '@/configs/dashboard.ts'
 import { ReturnSumData } from '@/types/components/dashboard/dashboard.ts'
 import { getBarCharts, getCountCards, getTableData } from '@/utils/dashboard/get-count-cards.ts'
